FIX: Giving up to clean modules for backend

+++ build.sh
@@ -2,12 +2,26 @@
 IMAGE="simple-todo"
 REPO="sokubedocker"
 
-for f in ./features/*; do
-    version=$(echo "$f" | cut -d '/' -f 3 )
-    echo "f=$f, version=$version"
-    cp "$f" features.json
-    docker build -t "$REPO/$IMAGE:$version" .
-    docker push "$REPO/$IMAGE:$version"
-done
+build_version () {
+    feature_file=$(echo "./features/$1")
+    if [ -f "$feature_file" ]; then
+        version=$1
+        echo "Building version $1 with features file=$feature_file"
+        cp "$feature_file" features.json
+        docker build -t "$REPO/$IMAGE:$version" .
+        docker push "$REPO/$IMAGE:$version"
+        rm -f features.json
+    else
+        echo "Version $1 does not exist in ./features/"
+    fi
+}
 
-rm -f features.json
+if [ ! -z "$1" ]; then
+    build_version $1
+else
+    for f in ./features/*; do
+        version=$(echo "$f" | cut -d '/' -f 3 )
+        build_version $version
+    done
+
+fi
